Revert "🤖 backported "Add Anchor Links to Documents" (#68519)"
This reverts commit 9699dfba7425e1d67078b2a6c6e003684b616285.
