+++ frontend/src/metabase/documents/components/DocumentPage.tsx
@@ -8,7 +8,6 @@ import {
   useCallback,
   useEffect,
   useMemo,
-  useRef,
   useState,
 } from "react";
 import type { Route } from "react-router";
@@ -63,7 +62,6 @@ import {
 } from "../documents.slice";
 import { useDocumentState } from "../hooks/use-document-state";
 import { useRegisterDocumentMetabotContext } from "../hooks/use-register-document-metabot-context";
-import { useScrollToAnchor } from "../hooks/use-scroll-to-anchor";
 import {
   getDraftCards,
   getHasUnsavedChanges,
@@ -102,7 +100,6 @@ export const DocumentPage = ({
   const [editorInstance, setEditorInstance] = useState<TiptapEditor | null>(
     null,
   );
-  const editorContainerRef = useRef<HTMLDivElement>(null);
   const hasUnsavedEditorChanges = useSelector(getHasUnsavedChanges);
   const [createDocument, { isLoading: isCreating }] =
     useCreateDocumentMutation();
@@ -203,14 +200,6 @@ export const DocumentPage = ({
   useEffect(() => {
     dispatch(setChildTargetId(paramsChildTargetId));
   }, [dispatch, paramsChildTargetId]);
-
-  // Scroll to anchor block when navigating with URL hash
-  const blockId = location.hash ? location.hash.slice(1) : null;
-  useScrollToAnchor({
-    blockId,
-    editorContainerRef,
-    isLoading: isDocumentLoading,
-  });
 
   const hasUnsavedChanges = useCallback(() => {
     const currentTitle = documentTitle.trim();
@@ -476,7 +465,6 @@ export const DocumentPage = ({
               onChange={handleChange}
               editable={canWrite}
               isLoading={isDocumentLoading}
-              editorContainerRef={editorContainerRef}
             />
           </Box>
         </Box>
