--- frontend/src/metabase/documents/components/Editor/AnchorLinkMenu/AnchorLinkMenu.module.css
@@ -1,13 +1,0 @@
-.anchorLinkMenu {
-  opacity: 0;
-  user-select: none;
-
-  &:hover,
-  &:focus-within {
-    opacity: 1;
-  }
-}
-
-.visible {
-  opacity: 1;
-}
