--- frontend/src/metabase/documents/components/Editor/AnchorLinkMenu/index.ts
@@ -1,1 +1,0 @@
-export * from "./AnchorLinkMenu";
