+++ frontend/src/metabase/rich_text_editing/tiptap/extensions/BulletList/BulletList.tsx
@@ -1,3 +1,4 @@
+import { autoUpdate, useFloating } from "@floating-ui/react";
 import type { NodeViewProps } from "@tiptap/core";
 import { BulletList } from "@tiptap/extension-bullet-list";
 import {
@@ -6,9 +7,20 @@ import {
   ReactNodeViewRenderer,
 } from "@tiptap/react";
 import cx from "classnames";
+import { useEffect, useMemo, useState } from "react";
 
+import { useListCommentsQuery } from "metabase/api";
+import { getTargetChildCommentThreads } from "metabase/comments/utils";
 import { CommentsMenu } from "metabase/documents/components/Editor/CommentsMenu";
-import { useBlockMenus } from "metabase/documents/hooks/use-block-menus";
+import {
+  getChildTargetId,
+  getCurrentDocument,
+  getHoveredChildTargetId,
+} from "metabase/documents/selectors";
+import { getListCommentsQuery } from "metabase/documents/utils/api";
+import { isTopLevel } from "metabase/documents/utils/editorNodeUtils";
+import { isWithinIframe } from "metabase/lib/dom";
+import { useSelector } from "metabase/lib/redux";
 
 import { createIdAttribute, createProseMirrorPlugin } from "../NodeIds";
 import S from "../extensions.module.css";
@@ -30,19 +42,34 @@ export const CustomBulletList = BulletList.extend({
 });
 
 export const BulletListNodeView = ({ node, editor, getPos }: NodeViewProps) => {
-  const {
-    _id,
-    isOpen,
-    isHovered,
-    hovered,
-    setHovered,
-    threads,
-    document,
-    shouldShowMenus,
-    setReferenceElement,
-    commentsRefs,
-    commentsFloatingStyles,
-  } = useBlockMenus({ node, editor, getPos });
+  const childTargetId = useSelector(getChildTargetId);
+  const hoveredChildTargetId = useSelector(getHoveredChildTargetId);
+  const document = useSelector(getCurrentDocument);
+  const { data: commentsData } = useListCommentsQuery(
+    getListCommentsQuery(document),
+  );
+  const comments = commentsData?.comments;
+  const [hovered, setHovered] = useState(false);
+  const [rendered, setRendered] = useState(false); // floating ui wrongly positions things without this
+  const { _id } = node.attrs;
+  const isOpen = childTargetId === _id;
+  const isHovered = hoveredChildTargetId === _id;
+  const threads = useMemo(
+    () => getTargetChildCommentThreads(comments, _id),
+    [comments, _id],
+  );
+  const { refs, floatingStyles } = useFloating({
+    placement: "right-start",
+    whileElementsMounted: autoUpdate,
+    strategy: "fixed",
+    open: rendered,
+  });
+
+  useEffect(() => {
+    if (!rendered) {
+      setRendered(true);
+    }
+  }, [rendered]);
 
   return (
     <>
@@ -51,8 +78,7 @@ export const BulletListNodeView = ({ node, editor, getPos }: NodeViewProps) => {
         className={cx(S.root, {
           [S.open]: isOpen || isHovered,
         })}
-        data-node-id={_id}
-        ref={setReferenceElement}
+        ref={refs.setReference}
         // onMouseEnter/onMouseLeave do not work on list elements living in contentEditable
         onMouseOver={() => setHovered(true)}
         onMouseOut={() => setHovered(false)}
@@ -60,16 +86,19 @@ export const BulletListNodeView = ({ node, editor, getPos }: NodeViewProps) => {
         <NodeViewContent<"ul"> as="ul" />
       </NodeViewWrapper>
 
-      {shouldShowMenus && document && (
-        <CommentsMenu
-          active={isOpen}
-          href={`/document/${document.id}/comments/${_id}`}
-          ref={commentsRefs.setFloating}
-          show={isOpen || hovered}
-          style={commentsFloatingStyles}
-          threads={threads}
-        />
-      )}
+      {document &&
+        rendered &&
+        isTopLevel({ editor, getPos }) &&
+        !isWithinIframe() && (
+          <CommentsMenu
+            active={isOpen}
+            href={`/document/${document.id}/comments/${_id}`}
+            ref={refs.setFloating}
+            show={isOpen || hovered}
+            threads={threads}
+            style={floatingStyles}
+          />
+        )}
     </>
   );
 };
