+++ frontend/src/metabase/rich_text_editing/tiptap/extensions/Heading/Heading.tsx
@@ -1,3 +1,4 @@
+import { autoUpdate, useFloating } from "@floating-ui/react";
 import type { NodeViewProps } from "@tiptap/core";
 import { Heading } from "@tiptap/extension-heading";
 import {
@@ -6,10 +7,20 @@ import {
   ReactNodeViewRenderer,
 } from "@tiptap/react";
 import cx from "classnames";
+import { useEffect, useMemo, useState } from "react";
 
-import { AnchorLinkMenu } from "metabase/documents/components/Editor/AnchorLinkMenu";
+import { useListCommentsQuery } from "metabase/api";
+import { getTargetChildCommentThreads } from "metabase/comments/utils";
 import { CommentsMenu } from "metabase/documents/components/Editor/CommentsMenu";
-import { useBlockMenus } from "metabase/documents/hooks/use-block-menus";
+import {
+  getChildTargetId,
+  getCurrentDocument,
+  getHoveredChildTargetId,
+} from "metabase/documents/selectors";
+import { getListCommentsQuery } from "metabase/documents/utils/api";
+import { isTopLevel } from "metabase/documents/utils/editorNodeUtils";
+import { isWithinIframe } from "metabase/lib/dom";
+import { useSelector } from "metabase/lib/redux";
 
 import { createIdAttribute, createProseMirrorPlugin } from "../NodeIds";
 import S from "../extensions.module.css";
@@ -48,24 +59,34 @@ const levelNodeMap: Record<Level, ElementType> = {
 };
 
 export const HeadingNodeView = ({ node, editor, getPos }: NodeViewProps) => {
-  const { level } = node.attrs;
+  const childTargetId = useSelector(getChildTargetId);
+  const hoveredChildTargetId = useSelector(getHoveredChildTargetId);
+  const document = useSelector(getCurrentDocument);
+  const { data: commentsData } = useListCommentsQuery(
+    getListCommentsQuery(document),
+  );
+  const comments = commentsData?.comments;
+  const [hovered, setHovered] = useState(false);
+  const [rendered, setRendered] = useState(false); // floating ui wrongly positions things without this
+  const { _id, level } = node.attrs;
+  const isOpen = childTargetId === _id;
+  const isHovered = hoveredChildTargetId === _id;
+  const threads = useMemo(
+    () => getTargetChildCommentThreads(comments, _id),
+    [comments, _id],
+  );
+  const { refs, floatingStyles } = useFloating({
+    placement: "right-start",
+    whileElementsMounted: autoUpdate,
+    strategy: "fixed",
+    open: rendered,
+  });
 
-  const {
-    _id,
-    isOpen,
-    isHovered,
-    hovered,
-    setHovered,
-    threads,
-    document,
-    shouldShowMenus,
-    anchorUrl,
-    setReferenceElement,
-    commentsRefs,
-    commentsFloatingStyles,
-    anchorRefs,
-    anchorFloatingStyles,
-  } = useBlockMenus({ node, editor, getPos });
+  useEffect(() => {
+    if (!rendered) {
+      setRendered(true);
+    }
+  }, [rendered]);
 
   return (
     <>
@@ -74,8 +95,7 @@ export const HeadingNodeView = ({ node, editor, getPos }: NodeViewProps) => {
         className={cx(S.root, {
           [S.open]: isOpen || isHovered,
         })}
-        data-node-id={_id}
-        ref={setReferenceElement}
+        ref={refs.setReference}
         onMouseOver={() => setHovered(true)}
         onMouseOut={() => setHovered(false)}
       >
@@ -84,24 +104,19 @@ export const HeadingNodeView = ({ node, editor, getPos }: NodeViewProps) => {
         />
       </NodeViewWrapper>
 
-      {shouldShowMenus && document && (
-        <>
-          <AnchorLinkMenu
-            ref={anchorRefs.setFloating}
-            show={hovered}
-            style={anchorFloatingStyles}
-            url={anchorUrl}
-          />
+      {document &&
+        rendered &&
+        isTopLevel({ editor, getPos }) &&
+        !isWithinIframe() && (
           <CommentsMenu
             active={isOpen}
             href={`/document/${document.id}/comments/${_id}`}
-            ref={commentsRefs.setFloating}
+            ref={refs.setFloating}
             show={isOpen || hovered}
-            style={commentsFloatingStyles}
             threads={threads}
+            style={floatingStyles}
           />
-        </>
-      )}
+        )}
     </>
   );
 };
