+++ frontend/src/metabase/rich_text_editing/tiptap/extensions/Paragraph/Paragraph.tsx
@@ -1,3 +1,4 @@
+import { autoUpdate, useFloating } from "@floating-ui/react";
 import type { NodeViewProps } from "@tiptap/core";
 import { Paragraph } from "@tiptap/extension-paragraph";
 import {
@@ -6,9 +7,20 @@ import {
   ReactNodeViewRenderer,
 } from "@tiptap/react";
 import cx from "classnames";
+import { useEffect, useMemo, useState } from "react";
 
+import { useListCommentsQuery } from "metabase/api";
+import { getTargetChildCommentThreads } from "metabase/comments/utils";
 import { CommentsMenu } from "metabase/documents/components/Editor/CommentsMenu";
-import { useBlockMenus } from "metabase/documents/hooks/use-block-menus";
+import {
+  getChildTargetId,
+  getCurrentDocument,
+  getHoveredChildTargetId,
+} from "metabase/documents/selectors";
+import { getListCommentsQuery } from "metabase/documents/utils/api";
+import { isTopLevel } from "metabase/documents/utils/editorNodeUtils";
+import { isWithinIframe } from "metabase/lib/dom";
+import { useSelector } from "metabase/lib/redux";
 
 import { createIdAttribute, createProseMirrorPlugin } from "../NodeIds";
 import S from "../extensions.module.css";
@@ -36,26 +48,36 @@ export const ParagraphNodeView = ({
   extension,
 }: NodeViewProps) => {
   const editorContext = extension?.options?.editorContext || "document";
-  const hideMenusInContext = editorContext === "comments";
-
-  const {
-    _id,
-    isOpen,
-    isHovered,
-    hovered,
-    setHovered,
-    threads,
-    document,
-    shouldShowMenus,
-    setReferenceElement,
-    commentsRefs,
-    commentsFloatingStyles,
-  } = useBlockMenus({
-    node,
-    editor,
-    getPos,
-    shouldHideMenus: hideMenusInContext,
+  const shouldHideCommentMenu =
+    editorContext === "comments" || isWithinIframe();
+  const childTargetId = useSelector(getChildTargetId);
+  const hoveredChildTargetId = useSelector(getHoveredChildTargetId);
+  const document = useSelector(getCurrentDocument);
+  const { data: commentsData } = useListCommentsQuery(
+    getListCommentsQuery(document),
+  );
+  const comments = commentsData?.comments;
+  const [hovered, setHovered] = useState(false);
+  const [rendered, setRendered] = useState(false); // floating ui wrongly positions things without this
+  const { _id } = node.attrs;
+  const isOpen = childTargetId === _id;
+  const isHovered = hoveredChildTargetId === _id;
+  const threads = useMemo(
+    () => getTargetChildCommentThreads(comments, _id),
+    [comments, _id],
+  );
+  const { refs, floatingStyles } = useFloating({
+    placement: "right-start",
+    whileElementsMounted: autoUpdate,
+    strategy: "fixed",
+    open: rendered,
   });
+
+  useEffect(() => {
+    if (!rendered) {
+      setRendered(true);
+    }
+  }, [rendered]);
 
   return (
     <>
@@ -64,24 +86,26 @@ export const ParagraphNodeView = ({
         className={cx(S.root, {
           [S.open]: isOpen || isHovered,
         })}
-        data-node-id={_id}
-        ref={setReferenceElement}
+        ref={refs.setReference}
         onMouseOver={() => setHovered(true)}
         onMouseOut={() => setHovered(false)}
       >
         <NodeViewContent<"p"> as="p" />
       </NodeViewWrapper>
 
-      {shouldShowMenus && document && (
-        <CommentsMenu
-          active={isOpen}
-          href={`/document/${document.id}/comments/${_id}`}
-          ref={commentsRefs.setFloating}
-          show={isOpen || hovered}
-          style={commentsFloatingStyles}
-          threads={threads}
-        />
-      )}
+      {document &&
+        rendered &&
+        !shouldHideCommentMenu &&
+        isTopLevel({ editor, getPos }) && (
+          <CommentsMenu
+            active={isOpen}
+            href={`/document/${document.id}/comments/${_id}`}
+            ref={refs.setFloating}
+            show={isOpen || hovered}
+            threads={threads}
+            style={floatingStyles}
+          />
+        )}
     </>
   );
 };
